requirements_sales.txt
fastapi==0.120.1
gradio==5.49.1
httpx==0.28.1
langchain_community==0.4.1
langchain_core==1.0.1
langchain_openai==1.0.1
librosa==0.11.0
litellm==1.79.0
numpy==2.3.4
openai==2.6.1
pandas==2.3.3
pydantic==2.12.3
pydub==0.25.1
pytest==8.4.2
python-dotenv==1.2.1
SQLAlchemy==2.0.44
underthesea==8.3.0
uvicorn==0.38.0
alembic==1.13.1  